+++ packages/theme_tailor_toolbox/example/lib/benchmark.dart
@@ -4,8 +4,7 @@ import 'dart:developer';
 
 import 'package:benchmark_harness/benchmark_harness.dart';
 import 'package:flutter/material.dart';
-import 'package:theme_tailor_toolbox/src/theme_tailor_encoder/color_encoder.dart';
-import 'package:theme_tailor_toolbox/src/theme_tailor_encoder/material_color_encoder.dart';
+import 'package:theme_tailor_toolbox/theme_tailor_toolbox.dart';
 
 class BenchmarkScoreEmitter extends ScoreEmitter {
   BenchmarkScoreEmitter();
@@ -20,7 +19,7 @@ class ColorEncoderBenchmark extends BenchmarkBase {
   ColorEncoderBenchmark()
       : super('Encoder: Color', emitter: BenchmarkScoreEmitter());
 
-  final ColorEncoder encoder = const ColorEncoder();
+  final encoder = colorEncoder;
 
   @override
   void run() {
@@ -32,7 +31,7 @@ class ColorEncoderNullableBenchmark extends BenchmarkBase {
   ColorEncoderNullableBenchmark()
       : super('Encoder: Color?', emitter: BenchmarkScoreEmitter());
 
-  final ColorNullEncoder encoder = const ColorNullEncoder();
+  final encoder = colorNullEncoder;
 
   @override
   void run() {
@@ -44,7 +43,7 @@ class MaterialColorEncoderBenchmark extends BenchmarkBase {
   MaterialColorEncoderBenchmark()
       : super('Encoder: MaterialColor', emitter: BenchmarkScoreEmitter());
 
-  final MaterialColorEncoder encoder = const MaterialColorEncoder();
+  final encoder = materialColorEncoder;
 
   @override
   void run() {
@@ -56,7 +55,7 @@ class MaterialColorEncoderNullableBenchmark extends BenchmarkBase {
   MaterialColorEncoderNullableBenchmark()
       : super('Encoder: MaterialColor?', emitter: BenchmarkScoreEmitter());
 
-  final MaterialColorNullEncoder encoder = const MaterialColorNullEncoder();
+  final encoder = materialColorNullEncoder;
 
   @override
   void run() {
@@ -68,7 +67,7 @@ class MaterialAccentColorEncoderBenchmark extends BenchmarkBase {
   MaterialAccentColorEncoderBenchmark()
       : super('Encoder: MaterialAccentColor', emitter: BenchmarkScoreEmitter());
 
-  final MaterialAccentColorEncoder encoder = const MaterialAccentColorEncoder();
+  final encoder = materialAccentColorEncoder;
 
   @override
   void run() {
@@ -81,8 +80,7 @@ class MaterialAccentColorEncoderNullableBenchmark extends BenchmarkBase {
       : super('Encoder: MaterialAccentColor?',
             emitter: BenchmarkScoreEmitter());
 
-  final MaterialAccentColorNullEncoder encoder =
-      const MaterialAccentColorNullEncoder();
+  final encoder = materialAccentColorNullEncoder;
 
   @override
   void run() {
