+++ packages/theme_tailor_toolbox/lib/src/theme_tailor_encoder/color/color_encoder.dart
@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:theme_tailor_annotation/theme_tailor_annotation.dart';
 
+import '../no_interpolation_encoder.dart';
+
 const colorEncoder = ColorEncoder();
-const colorNullEncoder = ColorNullEncoder();
+const colorEncoderNoLerp = ColorEncoderNoLerp();
 
 class ColorEncoder extends ThemeEncoder<Color> {
   const ColorEncoder();
@@ -11,9 +13,6 @@ class ColorEncoder extends ThemeEncoder<Color> {
   Color lerp(Color a, Color b, double t) => Color.lerp(a, b, t)!;
 }
 
-class ColorNullEncoder extends ThemeEncoder<Color?> {
-  const ColorNullEncoder();
-
-  @override
-  Color? lerp(Color? a, Color? b, double t) => Color.lerp(a, b, t);
+class ColorEncoderNoLerp extends NoLerpEncoder<Color> {
+  const ColorEncoderNoLerp();
 }
