+++ packages/theme_tailor_toolbox/lib/src/theme_tailor_encoder/color/util.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+Color colorAlphaScale(Color a, double factor) {
+  return a.withAlpha((a.alpha * factor).round().clamp(0, 255));
+}
